requirements.txt
alembic==1.7.5
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
zipp==3.6.0
